requirements.txt
Flask
line-bot-sdk
Gunicorn
